Refactor code structure
Extract push related code into separate module.
